Platform and second bar in VrScene

+++ engine/src/main/java/de/yard/threed/engine/FirstPersonController.java
@@ -137,7 +137,7 @@ public class FirstPersonController {
         this.movementSpeed = movementSpeed;
     }
 
-    public void setRotationSpeed(float rotationSpeed) {
+    public void setRotationSpeed(double rotationSpeed) {
         this.rotationSpeed = rotationSpeed;
     }
